Fix issue #34: テストの修正
